fixed chat service
