+++ NetworkApi/Sources/NetworkApi/Utils/SafeCodable.swift
@@ -33,3 +33,9 @@ public struct SafeCodable<T: Codable & UnknownSafable>: Codable {
 public protocol UnknownSafable {
     static var unknown: Self { get }
 }
+
+extension Optional: UnknownSafable where Wrapped: UnknownSafable {
+    public static var unknown: Self {
+        return nil
+    }
+}
