+++ Stroganina/Models/Mapping/Message+Raw.swift
@@ -9,14 +9,30 @@ import Foundation
 import NetworkApi
 
 extension Message {
-    convenience init(_ raw: Raw.Message) {
+    convenience init(_ raw: Raw.Message, identifier: MessageIdentifier) {
         self.init(
-            id: "\(raw.messageId)",
+            id: identifier,
             date: Date(timeIntervalSince1970: TimeInterval(raw.date)),
-            user: User(raw.user),
-            isOutgoing: raw.user.isSelf,
+            user: raw.user.flatMap { User($0) },
+            isOutgoing: raw.user?.isSelf ?? true,
             chatId: raw.chatId,
-            state: .unknown
+            remoteId: raw.messageId,
+            state: .init(raw.state?.value)
         )
+    }
+}
+
+extension Message.MessageState {
+    init(_ state: Raw.MessageState?) {
+        switch state {
+        case .watingSend:
+            self = .watingSend
+        case .none:
+            self = .sended
+        case .failed:
+            self = .failed
+        case .read:
+            self = .read
+        }
     }
 }
