+++ Stroganina/Services/UpdateCenter/Notification.swift
@@ -8,7 +8,7 @@
 import NetworkApi
 
 enum Notification {
-    case newMessage(_ message: Raw.Message)
-    case newChat(_ chat: Raw.Chat)
+    case newMessage(_ message: MessageWrapper)
+    case newChat(_ chat: Chat)
     case closeConnect
 }
